requirements.txt
keras==2.7.0
matplotlib==3.2.2
numpy==1.19.5
opencv_python==4.1.2.30
pandas==1.1.5
Pillow==8.4.0
seaborn==0.11.2
streamlit==1.2.0
tensorflow==2.7.0
